+++ packages/orm/src/orm.ts
@@ -4,6 +4,7 @@ import type { Executor } from './executor';
 export interface ORMConfig {
 	apis: Record<string, any>;
 	executor: Executor;
+	extraHeaders?: Record<string, string>;
 }
 
 export class ORM<Schemas extends Record<string, any>> {
@@ -16,6 +17,14 @@ export class ORM<Schemas extends Record<string, any>> {
 			schema: (this.config.apis as any)[namespace],
 			executor: this.config.executor,
 			namespace,
+			extraHeaders: this.config.extraHeaders,
+		});
+	}
+
+	withHeaders(headers: Record<string, string>): ORM<Schemas> {
+		return new ORM<Schemas>({
+			...this.config,
+			extraHeaders: headers,
 		});
 	}
 }