+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.6](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.5...@wundergraph/golang-client@0.6.6) (2023-06-05)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.6.5](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.4...@wundergraph/golang-client@0.6.5) (2023-06-04)
 
 **Note:** Version bump only for package @wundergraph/golang-client